code cleanup: rename oas3 -> openapi
