+++ src/dso_api/dynamic_api/routers.py
@@ -13,7 +13,7 @@ from schematools.utils import to_snake_case
 
 from dso_api.dynamic_api.app_config import register_model
 from dso_api.dynamic_api.locking import lock_for_writing
-from dso_api.dynamic_api.oas3 import get_openapi_json_view
+from dso_api.dynamic_api.openapi import get_openapi_json_view
 from dso_api.dynamic_api.remote import remote_serializer_factory, remote_viewset_factory
 from dso_api.dynamic_api.serializers import get_view_name, serializer_factory
 from dso_api.dynamic_api.views import viewset_factory
